requirements.txt
python-telegram-bot==20.5

transformers 
pandas 
torch
scikit-learn